add isclose function; used by math.eq
I finally got fed up with instances that look like `1 != 1`.
The left-hand side was the result of a calculation like `sqrt(x^2+y^2)`,
so some floating point error made the stored value slightly less than one. But it's displayed as '1'.

This adds a function `Numbas.math.isclose`, which tells if two numbers
are approximately equal, by default checking to a relative difference of
1e-15. It's based on [Python's PEP 485](https://www.python.org/dev/peps/pep-0485/#rationale).

Numbas.math.eq uses isclose to make decisions. The comparison operators
also use isclose to check for equality before checking for difference.
(So `a<b` if `not (a>=b)`.)

This also adds a JME function `isclose(a,b,rel_tol,abs_tol)`.