add `make_variables` JME function; JME part value generators are evaluated like question variables
`make_variables` calls Numbas.jme.variables.makeVariables, the routine
that evaluates a setof variable definitions with interdependencies.

So, given a dictionary mapping variable names to their definitions, and
those definitions can be in terms of other varibles in the dictionary,
evaluate them in the right order and return a dictionary of the values.

The value generator expressions for JME parts are evaluated using this,
so the value of one variable can be given in terms of others.