add 'render' JME function
Towards #423.

We already have the 'safe' property for strings, which prevents variable
substitution, so this adds the other end of the process:
`render(str,variables)` substitutes variables into the given string,
adding the given dictionary of variables into the scope temporarily.

`let(variables,render(str))` would be equivalent, but is much less
legible.