some fiddling with the pause/end buttons
see #580

the end exam button is now almost full width, and has red "danger"
styling.

The pause button has some margin below it, so it should be harder to
accidentally press the wrong button.

We spent a long time looking at different places to put the pause
button, or ways of styling it. I want to get rid of it, but students
need something to tell them it's safe to just close the window and
they can come back.

We considered a non-clickable "data saved" indicator, but because of the
way SCORM works it can't tell the truth when data is saved
asynchronously. And it doesn't solve the problem when you really want to
allow pausing.