type casting! More number types!
I think I've found a way of having more than one number type, and
automatically convertnig when required.

Tokens can now have a `casts` dictionary, containing functions which
convert tokens to specified types. Number-like types all have a 'number'
cast.

Rather than checking for token.type==x, in many cases we now have
to check if a token can be cast to type x, with jme.isType(x), and then
cast with jme.castToType(x).

There's some logic to find the 'closest' compatible type for two objects
in jme.findCompatibleType.

The parser now parses number literals as either 'integer' or 'number'
types, so that we can take advantage of knowing that a particular number
is an integer without explicitly casting, or checking for integer-ness
at each use.

The evaluation logic is more complicated, to account for implicit
casting. The `intype` property of funcObj is now a function which, given
a list of arguments, returns either `false` or a list giving a desired
type for each argument. The definition of a function which returns a
list of desired types closest to the actual types is picked, and any
casts are performed automatically before evaluation.

This means you can use a function defined only on 'number' (which uses
JS numbers) with any type that can implicitly cast to number'.

Division of integers produces a rational, so arithmetic operations on
integers should never lose accuracy (unless you exceed MAX_SAFE_INTEGER).

All the unit tests pass, so I need to write more. And then I need to
implement more number types.