Big decimals!
This commit adds the Decimal.js library to power a new 'decimal' data
type. Decimals offer much higher precision than JS numbers.

see #531

There's a new 'as' operator which tries to cast a value to the given
type, e.g. `1.2 as "decimal"`.

I'm now working on allowing funcObjs to specify the types of list
elements, with the aim of automatically casting number-like values to go
into things like the matrix constructor.
We'll be able to say things like "this parameter is a list of lists of
numbers".

Number entry parts use decimals for the min/maxvalue, which should make
it possible to mark very very large and very very small numbers.

An integer to the power of an integer produces a `decimal`, not a
`number`. This is a hacky way of ensuring that things like `6.022*10^23`
end up as decimals and don't lose precision.

I need to write lots and lots of unit tests.