more type-checking!
I've got rid of all custom typecheck methods in jme-builtins, and
replaced them with the new signature checking functions.

When comparing function definitions, an 'any type' signature counts as
an implicit cast.

Therer are 'list of' and 'dictionary of' signature checkers. I need to
make inferVariableTypes work with them. I also want an 'enum' signature,
which matches a dictionary with only the given keys, whose values have
the given types.