jme.inferVariableTypes works with casting
Lots and lots of plumbing to make it possible to find an assignment of
types to variables that will allow an expression to be evaluated.
Doesn't work with dependent types or with functions with variable
numbers of arguments. If those ever come up as answers to a question,
I'll be sad.