always show feedback to instructors
The Numbas LTI provider adds a SCORM element 'numbas.user_role', which
has the value 'instructor' when the exam is being viewed by an
instructor.

When viewed by an instructor, all feedback is visible, and the results
page is shown even if "show results page" is turned off.

This is mentioned in #589