jme.signature.optional produces a 'missing' item
Optional parameters for JME functions produce a 'missing' item in the
call signature, which produces the value `undefined` by default.

This way, it's easier to identify which arguments have been given. The
`evaluate` function is now also given a third argument `signature`,
which is the output of the typecheck function.

For example, if a function has the signature
`optional(string),optional(number)`, the calling patterns work as
follows:

* [] -> [undefined,undefined]
* [string] -> [string,undefined]
* [number] -> [undefined,number]
* [string,number] -> [string,number]