when adding in extras to a new scope, don't reuse the rulesets and functions dictionaries
When constructing a new scope from both a parent and an 'extras' scope,
add rulesets and functions to blank dictionaries created for the new
scope, instead of reusing the dictionaries from the 'extras' scope.

Reusing the dictionaries meant that the 'extras' scope was being
polluted with whatever was added to the new scope.

fixes #631