Starting on choose your own adventure mode!
There's a new mode for questions: 'adaptive parts'. Initially, just an
instance of the first part is generated.

Each part can define a list of "next parts" - options for what to do
next. They can replace the values of question variables with the values
of marking notes. The rest of the variables are re-evaluated, like in
adaptive marking.

The next part is generated using this new scope.