more work on "next parts"
Each Next Part can define a condition for its availability, in terms of
question variables, marking notes, and the credit awarded.

Only one part is displayed. Because you might want to take more than one
option, showing everything linearly would be confusing. There's a link
above the current part to go back to the previous part. One an option
has been taken, the generated part is kept forever, so clicking on that
option takes you back to the same instance of the part.

I'm not sure how to calculate the score for the question. At the moment,
it's the score and marks available from all parts leading from the
current part up to the first part, ignoring any other branches.
Some other options:
* Count all parts generated (might penalise a student for picking an
  option too early, backtracking, then picking that option again
  later)
* Only count the last part (hard to give credit for work done before
  that)
* Author can define named "score bins", with fixed maximums, and each
  part feeds into a particular bin. Some logic for combining bins
  (credit for partial work, always counted, etc.)