add adaptive marking penalty
There's now an option to reduce the number of marks available when
adaptive marking is used. The penalty works the same way as the steps
penalty: the total number of marks avaialable is reduced, and the
student's score is `credit * available marks`.

The steps penalty and adaptive marks penalty combine additively: they're
both subtracted from the maximum available marks before working out the
score.