objectives and penalties in adaptive parts mode
When a question is presented in adaptive mode, scores from parts are
collected into 'objectives'. Each objective has a limit, so you can't
run up a score by repeatedly visiting a part.
There are also penalties, which are applied when visiting parts. The
penalty is defined in the "next parts" list, so you can see in advance
how much visiting a part will cost you.

The question's score is calculated by adding up all the objectives and
subtracting all the penalties. The question has a static maximum number
of marks.