remove sameVars note from JME marking algorithm
fixes #632

This removes check in the JME part's marking algorithm that the student used exactly the same set of
variables as the correct answer.
It was only ever a time-saving shortcut - the assumption was that if the
expressions don't use the same variables, they can't be the "same".
This is false: for example, x = x + 0*y.

You might also want to give the student feedback based on their answer
even if it doesn't use the right variables, such as the pattern
restriction.