default objective for parts in explore mode; improve the breakdown table
If a part doesn't have an objective assigned, its marks go into a
"default" objective which contributes directly to the question total but
doesn't appear in the list of objectives.

The objective/penalty breakdown now appears in the bottom question nav
bar instead of the "total" line. Each objective gets an icon showing
correct/partial/incorrect state. There's a final line showing the total
score for the question.

(Need to check how this works with different feedback settings).