tidy up re_op in jme.Parser
The regex to parse an operator now definitely checks for longer
operators first - even the built-in operators are part of the list
Parser.ops, instead of being implicitly defined in the regex.

The regex is constructed by Parser.make_re, which is called when the
parser is constructed, and whenever an operator is added.

jme.Scope objects now have a `parser` property, which is the standard
parser by default.

There's a unit test demonstrating how to create a custom parser, add an
operator to it, and use it in a scope.