numberentry fraction correct answer is less precise when using "number" values
JS Numbers are only accurate to around 14 decimal places. In the past,
we've used 12 decimal places to account for accumulation of error.

The numberentry part `getCorrectAnswer` method uses Fraction.fromDecimal
to find a rational approximation to the correct answer. If the Decimal
value was obtained by converting from a JS Number, it'll look more
precise than it actually is, so something like 30/7 would end up with a
huge denominator of the form 10^n.

This commit changes the logic to only find an approximation correct to
12 decimal places when the min/max values are "number" values, not
"decimal".