jme part does its best to mark formulas
The marking script for mathematical expression parts now detects when
the correct answer is a formula (an expression of the form `name =
expression`), and tries to set up the variable value generator for the
left-hand side so it's the same type as the right-hand side.
This means that any equation representing the same statement as the
formula is marked correct.