Working on extracting part tests from the editor
Parts loaded from JSON now have a .json attribute storing the source
data.

The "you were awarded X marks" messages are now added to marking
feedback in the part, not by the theme (Need to make sure custom themes
overriding part-display.js are updated)

While debugging, I spotted that Numbas.checkAllScriptsLoaded was not
checking some files.

I've written a django management command to collect a list of questions
from the editor, which is then fed into the part unit testing page.
For each unit test defined on parts in the question, it checks that the
expected value, feedback messages, and credit are generated.