add a collectNumbers rule to deal with repeated division
The simplifier rewrites division `a/b` as `a` times a unary division
applied to `b`, so that multiplication over the numerators can work with
fewer rules.
collectNumbers was getting stuck in a loop rewriting `(1/10/10)*9`
because it was picking out the `1` and collecting it with the `9`, then
reintroducing another `1` on removing the unary division.

This commit adds a rule `a/b/c -> a/eval(b*c)` which is applied before
the `a*b -> eval(a*b)` rule, to prevent this.