fix math.unscientific on numbers with significand>=10
I didn't account for numbers like `310e6`, which aren't strictly in
scientific form but are in engineering form.

The 'scientific' number notation style accepts any number of digits in
the integer part (unless it starts with 0, in which case only 0 is
accepted), and then an optional decimal part followed by 'e' and an
exponent.