change "review" link to a button; remove href on questionSelector
If you middle-click a link with href="#", the same page opens in a new
tab. This can break SCORM sessions, since the exam opens without the
SCORM wrapper.

This commit changes the "review" link in the question breakdown to a
button, and removes the href attribute from the <a> tag in the question
selector (bootstrap needs this to be an <a> tag)