remove next parts if resubmitting parent would invalidate them
In explore mode, if the variable replacements for a next part depend on
the parent part's marking algorithm, changing your answer to the parent
part can make the next part invalid.

When resubmitting a part in such a situation, the student is now shown a
message warning that one or more next parts will be removed from the
question. If they confirm they want to submit, any next parts which
depend on anything from the marking algorithm are removed.