starting to implement alternative marking schemes
A part can have any number of 'alternative marking schemes', which are
copies of the part with different marking settings.

If the student doesn't get full credit using the part's main marking
settings, each of the alternatives is tried. If an alternative awards
more credit, its feedback is used instead, along with an optional extra
comment.