When an alternative is used, only give the main feedback message.
The marking feedback generated by an alternative might contain
misleading lines like "your answer is correct" if it uses an unmodified
marking algorithm.
So, if an alternative is used, just display the feedback message for
that alternative, and attach all the credit to that line. The other
feedback messages are discarded.