add "menu" question navigation mode
implements #673

In this mode, the student is shown a menu of all the questions in the
exam. They can click on a question's name to attempt it, and can come
back to the menu at any time.

The intention for this mode is to provide a bank of revision or
practice material that the student can dip into at will.

There's no "end exam" or "pause", because I can't think of a use case
for them.