save staged answers
When the student's staged (unsubmitted) answer to a part changes, it's
saved to a non-standard SCORM element 'cmi.interactions.N.staged_answer'.

The Numbas LTI tool will accept this, but strictly SCORM-compliant
players will reject it, so we can't count on it working. It's not saved
as part of the cmi.suspend_data element because that can become very
big, and we don't want to be sending and saving that any more than
necessary.

When the exam resumes, the last submitted answer is submitted, as
before, but then the staged answer is restored.

See #545 and #544.