explore mode and alternatives load from JSON
This commit makes the loadFromJSON methods correctly load questions in
explore mode and parts with alternative answers.

There are several unit tests in tests/parts.html for these features.

There's a new method, Part.markAlternatives, which returns information
about which alternative was used along with the raw marking script
result. The editor uses this in the 'marking algorithms' tab to show the
values and states of all marking notes, along with which alternative was
used.