question signal HTMLAttached only fires after part HTML generated
This adds an intermediate question signal, 'mainHTMLAttached', which
is triggered when the question HTML apart from the parts has been generated.
The parts then create their HTML, and 'HTMLAttached' is triggered once
all the parts have created their HTML.

This also adds a `signals` property to parts, and an 'HTMLAttached'
signal which is triggered when the part's HTML has been created.

fixes #711