improve texName
This was prompted by someone noticing that multiplication of
Greek symbols such as `theta * lambda` gets a multiplication
symbol in the middle, which isn't consistent with multiplication
of single Latin letters.

There's a fair bit of logic to do with displaying a name token, which
should only be done once. This has moved from jme.display to the
TName and TFunc constructors.

The logic for deciding whether to include a multiplication symbol now
looks at the `isLong` property of a name, rather than the length of the
string. A name is long if it would be rendered as more than one letter,
ignoring subscripts and primes.