dpformat and sigformat display as plain text in curly braces
in math mode instead of plain text.

dpformat and sigformat are marked as latex so that they can be
substituted easily into LaTeX, but should also be displayed as plain
text outside math mode.

This adds a new attribute, display_latex, to TString tokens, which says
that they should be rendered in math mode when substituted into the DOM
inside curly braces. Only the latex function sets this attribute.