improve expansion of juxtaposed variable names
jme.Scope.expandJuxtapositions used to just split long names up into
a series of names, each consisting of a single letter and then
subscripts or digits.

This doesn't account for Greek letters, which should be considered as a
single letter. It also had its own regular expression for names, which
could fall out of sync with the one in the parser.

This commit changes expandJuxtapositions to instead find the longest
left-prefix of the name which is considered by Numbas.jme.getNameInfo to
be a single letter, and then repeat on the remainder until the whole
name is consumed.

It also avoids multi-character subscripts which end in a letter, so for
example `m_xy` will be interpreted as `m_x*y` rather than `m_xy`.