exec() is lazy
Because of the way variable substitution works, you can create a
'function' token and substitute it in, but then `exec` needs to be
lazy so the token isn't evaluated.

exec now checks if the first argument has any args. If so, it evaluates
it and uses the resulting token. If not, it just uses the token.
Really, I don't think `exec` is particularly useful - substituting into
an expression works just as well and is easier to reason about.