Trigger partsBound when question parts have bound
When `HTMLAttached` triggers, each part's HTML is available via the
html property, but the question's html property does not contain the
individual parts because that section of the DOM has not finished binding.
This breaks backwards compatibility with older questions using the
`question-html-attached` signal and accessing `question.display.html`.

This commit adds a new `partsBound` signal which is triggered when the
DOM is has finished binding question parts by listening for the
`descendantsComplete` event from Knockout. The signal is then used to
trigger `question-html-attached` for backwards compatibility.