HTMLAttached is really only triggered once all the HTML is attached.
The HTMLAttached signals for questions and parts are now only triggered
when Knockout has finished attaching all the HTML to the page.

The descendantsComplete binding is called when all elements under the
element have finished binding. When those elements use the 'promise'
binding, it waits until the promise has resolved.

So each question's `HTMLAttached` signal is triggered once all the parts'
HTML has been attached to the page.

The 'mainHTMLAttached' signal is misnamed: it's triggered when the HTML promise
resolves, which doesn't necessarily mean that the HTML is in the page.

The Exam object waits for 'mainHTMLAttached' to initialise the display's
question list. Knockout then uses that to populate the page, leading to
bindings being applied and finally 'HTMLAttached' is triggered.