JS marking scripts overwrite mark_answer instead of mark
It's occurred to me that `mark_answer` is a lot less complicated than
`mark`, and is called by the JME marking function `mark_part`, so should
be the thing that is overwritten by a custom JS marking script.

This changes what a custom JS marking script should return. As far as I
know, nobody has written a JS marking script that returns something,
because it's undocumented, so I'm happy to make this breaking change.

For reference, a JS marking script should now return an object of type
Numbas.marking.marking_script_result, that is:

{
    states: a dictionary of lists of feedback items,
    values: a dictionary of JME values, containing at least the key 'interpreted_answer',
    state_valid: a dictionary of booleans, with at least the keys 'mark' and 'interpreted_answer',
    state_errors: a dictionary of Error objects
}

Fixes #726