add min/max on lists for each of the number-like types
min(list) and max(list) only had one implementation, which assumed the
members of the list were numbers.

This commit changes the signature of those functions to insist on lists
of TNum types, and adds implementations for integer, rational and
decimal so they return the right type rather than casting down to number.

There are other functions which use ordering on numbers, such as
`clamp`. Really there should be a 'traits' system for getting the right
comparison function.