IE fixes (!)
This uses xml.transformNode instead of $.xsl.transform (which was
supposed to be a compatibility wrapper!) on IE.

At some point in $.xsl.transform, IE replaces self-closing <input> tags with
unclosed tags, so then refuses to parse them. It's really HTML, so
xml.transformNode just returns a string and then we set the innerHTML of
a div to that string.

This also moves the <localise> handling to display.makeHTMLfromXML
instead of at the template-loading stage, so jQuery is operating on HTML
instead of XML. For some reason it decided it doesn't like doing that
any more.