make the entire runtime run in node
Working towards numbas/numbas-lti-provider#110.

This changes the script queueing stuff to add module.exports from each
script to the `window` and `global` objects, if they exist.

Now the entire Numbas runtime can run unmodified in node.js, if the
`browser-env` package is loaded first and you set `Numbas.display =
null` before trying to load an exam.

This commit also changes the compiler so it adds a file
'numbas-manifest.json' to the package, with the Numbas version and a
dict of features that could be added to over time, so a tool can
automatically work out if a package supports the feature it wants
without trying to run the JS.