inferVariableTypes: 'multiple' signature can't exceed the number of arguments
The variable type inference thing was getting stuck in a loop of adding
arguments to a function with arbitrarily many arguments. This makes it
stop once it's exceeded the number of arguments actually present.

The backtracking also seemed to be wrong: it was backtracking every
argument *after* the advanced argument, instead of before. I've changed
it, but I worry there was a reason it was that way round. All the unit
tests pass, so maybe it really was incorrect.