set staged answer when a part is answered
The staged answer for each part is saved whenever it changes, but
throttled not to happen more than once in 5 seconds. That means that if
the student submits an answer very quickly after entering their answer,
the staged answer is not equal to the submitted answer for up to 5
seconds, during which time the student might leave the exam.

This changes store.partAnswered to also save the staged answer, to
guarantee it's up to date with the submitted answer.