substitute variables into next part button labels
This makes sure that jme.contentsubvars runs over the labels on buttons
to choose the next part.

The button uses the parent part's scope, because the next part's scope
can't have been made yet, but the link shown in the breadcrumbs uses the
next part's scope, so there might be an inconsistency.

So this is some of the way to #799