some pattern-matching changes
The right-hand side of the `where operator is evaluated in a scope with
the captured names available.

When matching a function application, the arguments of the function have
the original options applied - they're not overridden with the
non-associative and non-commutative options used to match the arguments
as a term sequence.