don't unslash braces in latex
Until this commit, texify replaced \{ and \} in latex strings with { and
}.

I think this was a hacky way of escaping curly braces in JME strings,
before the safe() function existed. It made it very hard to produce the
latex '\{', so I've removed that logic.

This is a breaking change; I'll have to look for questions that produce
latex and see if any need changing.