actually, only latex marked as safe keeps \{
I found too many questions which rely on the slash-stripping, so I've
taken another tack.

The `safe` function can now take any expression instead of just a string
literal, and both `latex` and `safe` preserve each other's properties.

When texifying a string marked as both latex and safe, it's passed
through completely unmodified. If it's just marked as latex, \{ is
replaced with {.

So `latex(safe("\{"+1+"\}"))` produces the LaTeX code `\{1\}`