fix the order of answers in gapfills with adaptive marking
When the gaps are marked in a different order due to adaptive marking,
we need to read the student's answers for the corresponding parts!

The old marking script just read studentAnswer in the initial order.
This changes the marking script to follow the adaptive order, unless
"sort student's answers" is ticked.