user-controlled display options
This adds a "display options" button to the default theme. It launches
a modal dialog with controls for the page background colour, text
colour, and text size.

The default theme's CSS uses custom properties to apply these, so they
won't work in IE11, but there are fallbacks so at least the original
settings will be used.

It tries to automatically pick either black or white for shadows, trying to contrast
with the background colour.

fixes #797