always show "invalid" feedback
students have been missing that their answers are invalid, because when
"show answer state" is turned off, the only feedback that the answer is
invalid is the outline on the input box, and the popup if it's focused.

This makes it so that feedback lines marked as "invalid" will always be
shown.

For gap-fill parts, the headers aren't shown because they're "neutral",
so it could be hard to work out which gap an "invalid" feedback message
corresponds to. Maybe they should be marked as "heading", so they'll be
shown if there are any shown messages under them.