starting on adaptive testing
There's a new namespace, Numbas.adaptive, containing stuff to do with
knowledge graphs and a DiagnosticController, which runs 'adaptive
testing scripts'. These are like marking scripts, but run at the exam
level. There's one script at the moment, diagnosys.jme, which implements
(some of) the diagnosys algorithm.

The diagnosys knowledge graph is hardcoded in at the moment, and each
question's list of topics is pulled out of its tags.