exam-level feedback and actions after a question
fixes #811: diagnostic scripts have a "feedback" note which produces a
string to show to the student describing either the current state of an
ongoing attempt, or to sum up a completed attempt.

fixes #810: diagnostic scripts have a "next_actions" note which produces
a list of actions to offer to the student when they click the "Move to
the next question" button, as well as some feedback text which
might summarise the current state. Each action has a text label as well
as a new state to use and the next question to show if it's picked.

The `topics` dictionary contains a key "questions" which is a list of
questions within that topic. Those objects should be considered opaque,
but they're really dictionaries with the topic name and the number of
the question in the corresponding question group.