starting on redoing constants
see #691

The JME scope now keeps track of constants. Constant values aren't
subbed in during parsing, but on evaluation. This lets you map over a
name that's otherwise used as a constant.

The hard part is display: there are a lot of special cases in the
display code to do with pi, i and e. First of all, the display code now
needs to be given a scope. Then, I suppose each of the special cases
need to find out what the name of each of those special numbers is, if
they're defined at all. That might be slow. I suppose the scope could
make a note whenever a constant is set with one of the magic values.