starting to refactor Numbas.jme.display
There are too many functions passing around a settings object, and with
scope-defined constants they're going to need a scope too.

This commit refactors Numbas.jme.display.texify so that it creates an
instance of a Texifier class, which has `settings` and `scope`
attributes.

All the texifying routines are methods of this object.

The logic about when to include multiplication symbols and brackets
really depends on how tokens are displayed, as numerals or letters or in
terms of basis constants. I'm not sure if the internal methods of
Texifier should return a semantic hint along with the TeX code, or if I
should continue writing methods that say "when this is texified, what
will it look like?"