starting to work on hacking pi and i out of jme.display
Texifier now has a dictionary common_constants, which contains
information about the constants we have special cases for. It looks
through the constants defined in the scope, and picks out constants
equivalent to the ones we're interested in.

For pi, it picks out anything that is an integer multiple of pi, using
math.piDegree. That lets it recognise tau = 2pi. Then texNumber will use
that symbol, and scaling factor, to show multiples of the circle
constant.

For the imaginary unit, it looks for sqrt(-1) exactly. I'm not aware of
any widely-used convention that uses a different pair of basis numbers
than 1,sqrt(-1) for complex numbers.

For vectors, it tries to pick out the standard basis vectors: 1 in one
component, and 0 in all the others.

At the moment, only the circle constant is used, and only in texNumber.
I need to look at whether any of the other uses of piDegree look at the
scaling factor.