finish removing magic constants from texify
texify no longer refers explicitly to pi, e or i.

There were a couple of rules to do with e and the multiplication symbol,
which I think are really about having any letter. I've replaced these
with checks that the TeX for the argument starts with a digit.

I think if I was starting again, the texifier would do a lot of
rewriting, producing a structure that explicitly says when it looks like
brackets or letters or whatever, and then the rules are both easier to
state and easier to apply.

The complex-number parts of texRationalNumber and texRealNumber were
identical, so I've moved those to texNumber, and then the first two
functions can just concern themselves with real numbers.

I had a go at showing vectors in terms of standard unit vectors. The big
problem is that it ends up looking like a sum of terms, so if you want
to avoid always putting brackets it, the multiplication needs to know
what it looks like. I think it's best implemented as a simplification
rule which rewrites the tree, before it gets texified.