treeToJME is aware of constants
There's a little bit of repeated logic in the texify and treeToJME
methods, so I've made an abstract class Numbas.jme.display.JMEDisplayer.

treeToJME is aware of constants, and does the right things with pi, i
and e.

The pattern-matcher seems to be OK with constants, and the calculus
simplification rules seem to work.

I've changed the `thing` parameter in the texify functions to `tree`, to
match treeToJME