load custom constants in a question
Questions now load a list of custom constants, and built-in constants to
disable. These apply to the question's scope.

I've made more calls to the jme.display methods use a scope. I need to
do a final sweep to catch them all.

The built-in simplification rules have the explicit values of i and pi
subbed in, so that they don't have to be defined in the scope passed to
the matcher.

Still to do:

    * math.niceNumber is still outstanding: it needs to know the symbols for
      i, pi and e.
    * I need to write some unit tests for how constants should behave.
    * Check that SCORM resume still works.