add "interpreted as" before the JME answer preview
this adds the screenreader-only text "interpreted as" to the start of
the JME preview box. Both the box and the MathJax part are marked as
aria-live="polite" - need to test with a screenreader to establish if
this is right. I don't think it's helpful to keep reading out
"interpreted as" each time the maths changes, but it's needed when
reading through the page, to separate it from the input box.

see #831