fixes to the DIAGNOSYS algorithm
The list of topics for picking the next topic uses the state *after*
asnwering the current question.

It first tries to follow a leads to/depends on link from the current
topic. If none, it moves to the middle unknown topic. The weighting
could be changed: earlier for students likely to know less, later for
students likely to know more.