trying not to lose accuracy when converting numbers to decimals
The cast method for number to decimal aims for 15 decimal places instead
of 14.

When working out the fraction to display as the correct answer to a
number entry part when the min/max values are numbers instead of
decimals, math.rationalApproximation is used instead of the
Decimal.toFraction method, since the decimal claims to be more accurate
than it really is.