use DOMcontentsubvars on distractor messages
There was a question where one of the distractors had an img tag with
latex in the attribute, which caused contentsubvars to interpret the
curly braces as JME substitution.

This changes to use DOMcontentsubvars, so the whole thing is parsed as
HTML, then has stuff subbed in to that, then serialised to HTML again.