resume parts in explore mode in the right order
Resuming explore mode questions didn't work properly if the student
moved to a different existing part before choosing a "next part" option.
I'm surprised I didn't test this!

This commit changes the way that explore mode questions are resumed:
rather than starting at the root part and resuming all its next part
options in the order they're defined, the parts are now resumed in the
order they were created.

The part suspend data contains both the index of the part in its
definition list and the path of the previous part, which is enough for
Question.resume to work out what to do.