run subvars on the image() function; DOMcontentsubvars can return a different element
jme.variables.DOMcontentsubvars swaps <img> tags pointing to SVG files
with <object> tags, so that variables can be subbed into it.

The `html` JME function already calls DOMcontentsubvars on its output,
but the `image` function wasn't - this commit makes it do that.

When only an <img> tag is passed into DOMcontentsubvars, a different
element needs to be returned; DOMcontentsubvars now returns an element,
which might not be the one passed in.