pre-submit tasks for parts and custom answer widgets
This commit adds pre-submit tasks for parts: asynchronous tasks that
must be completed before the part can be marked.

When a part is submitted, the marking routine gets as far as
`mark_answer`, where it's just about to run the JME marking algorithm.
If there is a `pre_submit` note in the algorithm, it is evaluated,
producing a list of TPromise tokens. If the list is non-empty, then the
marking aborts for now, and the part is marked as 'waiting for
pre-submit tasks'.
Once all the promises have resolved, the results are cached,
and the part is submitted again. This time, the cache is used, and the
results are made available to the marking algorithm in a variable called
`pre_submit`.

A pre-submit task should produce a promise which resolves to a dictionary of JME tokens.

Extensions can define new input widgets. A widget definition consists
of:

* a name
* a constructor function for the widget
* the signature of the type of JME value that the widget produces
* definitions of the options the widget accepts (in the same format as
  settings for custom part types)
* a function to convert a raw value produced by the widget to a JME token

The widget is given a function to call when its value changes, and must
provide three methods:

* setAnswerJSON - fill in the widget with a given answer
* disable - make the widget uneditable
* enable - make the widget editable