don't clip questions in print view
fixes #901

The change to the `.invisible` class in d2f22123cb7085f73290fb0efe474e2d92a6d6e4
set a max-height: 100vh on invisible elements.

Questions have the .invisible class when not shown. This commit adds a
`print-visible` class, so that they're not invisible in print mode.

I think this is more robust than toggling the class when you click the
"print this results summary" button - the user might just use the
browser's print function, so everything needs to be displayed properly
all the time.