always make Question.variablesTodo
76a7211865e564ff3b5a4d8fa2fc906b77879994 changed question creation
so that variable definitions are loaded into `q.variableDefinitions`,
and `q.variablesTodo` is only created after functions have been made.

But this only happened when calling to generate variables, which doesn't
happen if resuming an attempt.

This commit splits out the code to make `q.variablesTodo`, so it's
always made.